ranking 앱 개발 완료. 당일에 서버를 처음으로 실행 시 새로운 데이터를 크롤링하여 적용함. (프론트는 X)
